Improve integration spec
